requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu118
numpy==2.2.6
opencv-python==4.13.0.92
pillow==12.1.1
torch==2.10.0
torchvision==0.25.0
torchaudio
flwr==1.27.0
scikit-learn==1.7.2
matplotlib==3.10.8
pandas==2.3.3
streamlit==1.55.0
tqdm==4.67.3
albumentations==2.0.8
seaborn==0.13.2